+++ src/pages/Hamburger/Setting/Setting.js
@@ -16,15 +16,13 @@ class Setting extends React.Component {
   LogOut = () => {
     SInfo.setItem('AutoToken', 'no', {});
 
-    networks
-      .delete('https://api.oboonmobility.com/v0/members/logout')
-      .then(res => {
-        if (res.data.success === true || res.data.success === 'true') {
-          removeHeader();
-          this.props.memberReset();
-          this.setState({ logoutSuccess: true });
-        }
-      });
+    networks.delete('/members/logout').then(res => {
+      if (res.data.success === true || res.data.success === 'true') {
+        removeHeader();
+        this.props.memberReset();
+        this.setState({ logoutSuccess: true });
+      }
+    });
   };
 
   render() {
