+++ src/theme/size.js
@@ -1,4 +1,4 @@
 import { Dimensions } from 'react-native';
 
 export const { width, height } = Dimensions.get('window');
-export const MARGIN= 24;
+export const MARGIN = 30;
